rearrange options files
