+++ webpack.config.js
@@ -1,11 +1,12 @@
 const path = require('path');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
   entry: {
     background: './src/background',
-    options: './src/options',
+    options: './src/lib/options/options',
     tongwen: './src/lib/tongwen/tongwen'
   },
   output: {
@@ -17,10 +18,14 @@ module.exports = {
     new CopyWebpackPlugin([
       { from: './src/LICENSE' },
       { from: './src/manifest.json' },
-      { from: './src/options.css' },
-      { from: './src/options.html' },
+      { from: './src/lib/options/options.css' },
       { from: './src/_locales', to: '_locales' },
       { from: './src/icon', to: 'icon' }
-    ])
+    ]),
+    new HtmlWebpackPlugin({
+      filename: 'options.html',
+      template: 'src/lib/options/options.html',
+      chunks: ['options']
+    })
   ]
 };
